+++ app/sprinkles/core/src/UserFrosting.php
@@ -16,7 +16,6 @@ use Slim\App;
 use Slim\Container;
 use UserFrosting\Sprinkle\Core\Sprinkle\SprinkleManager;
 use UserFrosting\Support\Exception\FileNotFoundException;
-use UserFrosting\UniformResourceLocator\ResourceLocator;
 
 /**
  * UserFrosting Main Class.
@@ -65,7 +64,7 @@ class UserFrosting
      */
     public function fireEvent($eventName, Event $event = null)
     {
-        /** @var EventDispatcher $events */
+        /** @var EventDispatcher $eventDispatcher */
         $eventDispatcher = $this->ci->eventDispatcher;
 
         return $eventDispatcher->dispatch($eventName, $event);
@@ -76,7 +75,7 @@ class UserFrosting
      *
      * @return App
      */
-    public function getApp()
+    public function getApp(): App
     {
         return $this->app;
     }
@@ -86,7 +85,7 @@ class UserFrosting
      *
      * @return Container
      */
-    public function getContainer()
+    public function getContainer(): Container
     {
         return $this->ci;
     }
@@ -94,7 +93,7 @@ class UserFrosting
     /**
      * Initialize the application.  Set up Sprinkles and the Slim app, define routes, register global middleware, and run Slim.
      */
-    public function run()
+    public function run(): void
     {
         $this->app->run();
     }
@@ -102,12 +101,10 @@ class UserFrosting
     /**
      * Register system services, load all sprinkles, and add their resources and services.
      */
-    protected function setupSprinkles()
+    protected function setupSprinkles(): void
     {
-        // Register basic sevices
-        $this->setupBasicServices();
-
         // Boot the Sprinkle manager, which creates Sprinkle classes and subscribes them to the event dispatcher
+        /** @var SprinkleManager */
         $sprinkleManager = $this->ci->sprinkleManager;
 
         try {
@@ -122,12 +119,10 @@ class UserFrosting
 
         $this->fireEvent('onSprinklesInitialized');
 
-        // Add Sprinkle resources (assets, templates, etc) to locator
-        $sprinkleManager->addResources();
+        /**
+         * @deprecated 4.5.0 Use `onSprinklesInitialized` event instead
+         */
         $this->fireEvent('onSprinklesAddResources');
-
-        // Register Sprinkle services
-        $sprinkleManager->registerAllServices();
         $this->fireEvent('onSprinklesRegisterServices');
     }
 
@@ -136,17 +131,6 @@ class UserFrosting
      */
     protected function setupBasicServices(): void
     {
-        /*
-         * Path/file locator service.
-         *
-         * Register custom streams for the application, and add paths for app-level streams.
-         *
-         * @return \UserFrosting\UniformResourceLocator\ResourceLocator
-         */
-        $this->ci['locator'] = function ($c) {
-            return new ResourceLocator(\UserFrosting\ROOT_DIR);
-        };
-
         /*
          * Set up sprinkle manager service.
          *
@@ -169,8 +153,11 @@ class UserFrosting
     /**
      * Setup UserFrosting App, load sprinkles, load routes, etc.
      */
-    protected function setupApp()
+    protected function setupApp(): void
     {
+        // Register basic sevices
+        $this->setupBasicServices();
+
         // Setup sprinkles
         $this->setupSprinkles();
 
@@ -193,7 +180,7 @@ class UserFrosting
      *
      * @param string $errorMessage Message to display [Default ""]
      */
-    protected function renderSprinkleErrorPage($errorMessage = '')
+    protected function renderSprinkleErrorPage(string $errorMessage = '')
     {
         ob_clean();
         $title = 'UserFrosting Application Error';
@@ -217,7 +204,7 @@ class UserFrosting
      *
      * @param string $errorMessage Message to display [Default ""]
      */
-    protected function renderSprinkleErrorCli($errorMessage = '')
+    protected function renderSprinkleErrorCli(string $errorMessage = '')
     {
         exit($errorMessage . PHP_EOL);
     }