+++ app/sprinkles/core/tests/Unit/Sprinkle/SprinkleManagerTest.php
@@ -12,6 +12,7 @@ namespace UserFrosting\Sprinkle\Core\Tests\Unit\Sprinkle;
 
 use Psr\Container\ContainerInterface;
 use Mockery as m;
+use UserFrosting\Sprinkle\Core\Sprinkle\Sprinkle;
 use UserFrosting\Sprinkle\Core\Tests\TestCase;
 use UserFrosting\Sprinkle\Core\Sprinkle\SprinkleManager;
 
@@ -20,6 +21,9 @@ class SprinkleManagerTest extends TestCase
     /** @var ContainerInterface $fakeCi Our mocked CI used for testing */
     protected $fakeCi;
 
+    /** @var string */
+    protected $basePath = __DIR__ . '/data/';
+
     public function setUp(): void
     {
         // We don't call parent function to cancel CI creation and get accurate test coverage
@@ -27,9 +31,6 @@ class SprinkleManagerTest extends TestCase
         $this->fakeCi = m::mock(ContainerInterface::class);
         $this->fakeCi->eventDispatcher = new eventDispatcherStub();
         $this->fakeCi->locator = new ResourceLocatorStub();
-
-        // Setup test sprinkle mock class so it can be found by `class_exist`
-        m::mock('UserFrosting\Sprinkle\Test\Test');
     }
 
     public function tearDown(): void
@@ -41,7 +42,7 @@ class SprinkleManagerTest extends TestCase
     /**
      * @return SprinkleManager
      */
-    public function testConstructor()
+    public function testConstructor(): SprinkleManager
     {
         $sprinkleManager = new SprinkleManager($this->fakeCi);
         $this->assertInstanceOf(SprinkleManager::class, $sprinkleManager);
@@ -53,7 +54,7 @@ class SprinkleManagerTest extends TestCase
      * @depends testConstructor
      * @param SprinkleManager $sprinkleManager
      */
-    public function testGetSetSprinklesPath(SprinkleManager $sprinkleManager)
+    public function testGetSetSprinklesPath(SprinkleManager $sprinkleManager): void
     {
         $sprinkleManager->setSprinklesPath('/foo');
         $this->assertSame('/foo', $sprinkleManager->getSprinklesPath());
@@ -62,7 +63,7 @@ class SprinkleManagerTest extends TestCase
     /**
      * @depends testConstructor
      */
-    public function testLoadSprinkleWithNonExistingFile()
+    public function testLoadSprinkleWithNonExistingFile(): void
     {
         $sprinkleManager = new SprinkleManager($this->fakeCi);
         $this->expectException(\UserFrosting\Support\Exception\FileNotFoundException::class);
@@ -70,19 +71,40 @@ class SprinkleManagerTest extends TestCase
     }
 
     /**
+     * Will test the instanceof sprinkle check is done + that the right class
+     * is gnerated with the exception message assertion
+     *
      * @depends testConstructor
-     * @param  SprinkleManager $sprinkleManager
+     * @param SprinkleManager $sprinkleManager
+     */
+    public function testGetSprinklesWihoutMockClass(SprinkleManager $sprinkleManager): void
+    {
+        // Setup test sprinkle mock class so it can be found by `class_exist`
+        class_alias(BlahSprinkleStub::class, 'UserFrosting\Sprinkle\Blah\Blah');
+
+        //$this->expectException(); //TODO
+        $this->expectExceptionMessage("UserFrosting\Sprinkle\Blah\Blah must be an instance of UserFrosting\Sprinkle\Core\Sprinkle\Sprinkle");
+        $sprinkleManager->bootSprinkle('blah');
+    }
+
+    /**
+     * @depends testConstructor
      * @return SprinkleManager
      */
-    public function testGetSprinkles(SprinkleManager $sprinkleManager)
+    public function testGetSprinkles(SprinkleManager $sprinkleManager): SprinkleManager
     {
+        // Setup test sprinkle class alias so it can be found by `class_exist`
+        class_alias(TestSprinkleStub::class, 'UserFrosting\Sprinkle\Test\Test');
+
+        $sprinkleManager->setSprinklesPath($this->basePath);
         $sprinkleManager->initFromSchema(__DIR__ . '/data/sprinkles.json');
+
         $sprinkles = $sprinkleManager->getSprinkles();
-        $this->assertEquals([
-            'foo'  => null,
-            'bar'  => null,
-            'test' => new \UserFrosting\Sprinkle\Test\Test(),
-        ], $sprinkles);
+        $this->assertIsArray($sprinkles);
+        $this->assertCount(3, $sprinkles);
+        $this->assertNull($sprinkles['foo']);
+        $this->assertNull($sprinkles['bar']);
+        $this->assertInstanceOf(Sprinkle::class, $sprinkles['test']);
 
         return $sprinkleManager;
     }
@@ -91,10 +113,24 @@ class SprinkleManagerTest extends TestCase
      * @depends testGetSprinkles
      * @param SprinkleManager $sprinkleManager
      */
-    public function testGetSprinkleNames(SprinkleManager $sprinkleManager)
+    public function testGetSprinkleNames(SprinkleManager $sprinkleManager): void
     {
         $sprinkles = $sprinkleManager->getSprinkleNames();
         $this->assertSame(['foo', 'bar', 'test'], $sprinkles);
+    }
+
+    /**
+     * @depends testGetSprinkles
+     * @param SprinkleManager $sprinkleManager
+     */
+    public function testGetSprinkle(SprinkleManager $sprinkleManager): void
+    {
+        $this->assertNull($sprinkleManager->getSprinkle('foo'));
+        $this->assertNull($sprinkleManager->getSprinkle('bar'));
+        $this->assertInstanceOf(Sprinkle::class, $sprinkleManager->getSprinkle('test'));
+
+        $this->expectExceptionMessage('Sprinkle fOo not found.');
+        $sprinkleManager->getSprinkle('fOo');
     }
 
     /**
@@ -106,10 +142,10 @@ class SprinkleManagerTest extends TestCase
      * @testWith        ["bar", true]
      *                  ["test", true]
      *                  ["foo", true]
-     *                  ["fOo", true]
-     *                  ["foO", true]
-     *                  ["Foo", true]
-     *                  ["FOO", true]
+     *                  ["fOo", false]
+     *                  ["foO", false]
+     *                  ["Foo", false]
+     *                  ["FOO", false]
      *                  ["fo", false]
      *                  ["o", false]
      *                  ["oo", false]
@@ -119,7 +155,7 @@ class SprinkleManagerTest extends TestCase
      *                  ["barfoo", false]
      *                  ["blah", false]
      */
-    public function testIsAvailable($sprinkleName, $isAvailable, SprinkleManager $sprinkleManager)
+    public function testIsAvailable($sprinkleName, $isAvailable, SprinkleManager $sprinkleManager): void
     {
         $this->assertSame($isAvailable, $sprinkleManager->isAvailable($sprinkleName));
     }
@@ -132,11 +168,9 @@ class SprinkleManagerTest extends TestCase
      *                  ["bar"]
      *                  ["test"]
      */
-    public function testGetSprinklePath($sprinkleName, SprinkleManager $sprinkleManager)
+    public function testGetSprinklePath($sprinkleName, SprinkleManager $sprinkleManager): void
     {
-        $basePath = __DIR__ . '/data/';
-        $sprinkleManager->setSprinklesPath($basePath);
-        $this->assertSame($basePath . $sprinkleName, $sprinkleManager->getSprinklePath($sprinkleName));
+        $this->assertSame($this->basePath . $sprinkleName, $sprinkleManager->getSprinklePath($sprinkleName));
     }
 
     /**
@@ -144,10 +178,9 @@ class SprinkleManagerTest extends TestCase
      * @depends testGetSprinklePath
      * @param SprinkleManager $sprinkleManager
      */
-    public function testGetSprinklePathWherePathDoesntExist(SprinkleManager $sprinkleManager)
+    public function testGetSprinklePathWherePathDoesntExist(SprinkleManager $sprinkleManager): void
     {
-        $basePath = __DIR__ . '/foo/';
-        $sprinkleManager->setSprinklesPath($basePath);
+        $sprinkleManager->setSprinklesPath(__DIR__ . '/foo/');
 
         $this->expectException(\UserFrosting\Support\Exception\FileNotFoundException::class);
         $sprinkleManager->getSprinklePath('foo');
@@ -158,7 +191,7 @@ class SprinkleManagerTest extends TestCase
      * @depends testGetSprinklePath
      * @param SprinkleManager $sprinkleManager
      */
-    public function testGetSprinklePathWhereSprinkleDoesntExist(SprinkleManager $sprinkleManager)
+    public function testGetSprinklePathWhereSprinkleDoesntExist(SprinkleManager $sprinkleManager): void
     {
         $this->expectException(\UserFrosting\Support\Exception\FileNotFoundException::class);
         $sprinkleManager->getSprinklePath('blah');
@@ -168,17 +201,17 @@ class SprinkleManagerTest extends TestCase
      * @depends testGetSprinkles
      * @param SprinkleManager $sprinkleManager
      */
-    public function testRegisterAllServices(SprinkleManager $sprinkleManager)
+    public function testRegisterServices(SprinkleManager $sprinkleManager): void
     {
         // Set Expectations for test sprinkle ServiceProvider
         // @see https://stackoverflow.com/a/13390001/445757
         $this->getMockBuilder('nonexistant')
-        ->setMockClassName('foo')
+        ->setMockClassName('FooService') // MockClassName doesn't accept namespace
         ->setMethods(['register'])
         ->getMock();
-        class_alias('foo', 'UserFrosting\Sprinkle\Test\ServicesProvider\ServicesProvider');
+        class_alias('FooService', 'UserFrosting\Sprinkle\Test\ServicesProvider\ServicesProvider');
 
-        $this->assertNull($sprinkleManager->registerAllServices());
+        $this->assertNull($sprinkleManager->registerServices('foo')); // TODO Assert operation of mock class
     }
 
     /**
@@ -186,12 +219,10 @@ class SprinkleManagerTest extends TestCase
      * @depends testGetSprinklePath
      * @param SprinkleManager $sprinkleManager
      */
-    public function testAddResources(SprinkleManager $sprinkleManager)
+    public function testAddResources(SprinkleManager $sprinkleManager): void
     {
-        $basePath = __DIR__ . '/data/';
-        $sprinkleManager->setSprinklesPath($basePath);
-
-        $this->assertNull($sprinkleManager->addResources());
+        $sprinkleManager->setSprinklesPath($this->basePath);
+        $this->assertNull($sprinkleManager->addSprinkleResources('foo')); //TODO Assert mock operation on locator
     }
 
     /**
@@ -200,9 +231,10 @@ class SprinkleManagerTest extends TestCase
      * @depends testConstructor
      * @depends testGetSprinkles
      */
-    public function testLoadSprinkleWithTxtFile()
+    public function testLoadSprinkleWithTxtFile(): void
     {
         $sprinkleManager = new SprinkleManager($this->fakeCi);
+        $sprinkleManager->setSprinklesPath($this->basePath);
         $sprinkleManager->initFromSchema(__DIR__ . '/data/sprinkles.txt');
         $this->assertCount(3, $sprinkleManager->getSprinkles());
     }
@@ -210,7 +242,7 @@ class SprinkleManagerTest extends TestCase
     /**
      * @depends testConstructor
      */
-    public function testLoadSprinkleWithBadJson()
+    public function testLoadSprinkleWithBadJson(): void
     {
         $sprinkleManager = new SprinkleManager($this->fakeCi);
         $this->expectException(\UserFrosting\Support\Exception\JsonException::class);
@@ -221,9 +253,10 @@ class SprinkleManagerTest extends TestCase
      * @depends testConstructor
      * @depends testIsAvailable
      */
-    public function testLoadSprinkleWithDuplicateSprinkles()
+    public function testLoadSprinkleWithDuplicateSprinkles(): void
     {
         $sprinkleManager = new SprinkleManager($this->fakeCi);
+        $sprinkleManager->setSprinklesPath($this->basePath);
         $sprinkleManager->initFromSchema(__DIR__ . '/data/sprinkles-duplicate.json');
         $this->assertEquals([
             'foo'  => null,
@@ -231,7 +264,9 @@ class SprinkleManagerTest extends TestCase
             'bar'  => null,
         ], $sprinkleManager->getSprinkles());
 
-        $this->assertTrue($sprinkleManager->isAvailable('Foo'));
+        $this->assertFalse($sprinkleManager->isAvailable('Foo'));
+        $this->assertTrue($sprinkleManager->isAvailable('foo'));
+        $this->assertTrue($sprinkleManager->isAvailable('FOO'));
     }
 }
 
@@ -247,4 +282,12 @@ class ResourceLocatorStub
     public function registerLocation()
     {
     }
+}
+
+class TestSprinkleStub extends Sprinkle
+{
+}
+
+class BlahSprinkleStub
+{
 }